Visual indicator in MeetupDetailsView indicating that you have successfully RSVPd, animating into a green checkmark which persists if the user is still either pending or accepted.
TabView now has a visual indication when a user has received a message, and the conversation in MessagesView will be bolded with a blue dot to indicate the conversation is unread. Needs a fix where the TabView updates in real time but I will worry about that later.

Made a few sweeping adjustments to actually using my coding keys because I was being stupid and hardcoding and it caused me some headaches